Added export_sis and SIS handler to 30-scaler Added xspress3 handler to 31-xspress3 Added flyer class definitions to 32-zebra Added export_zebra to 32-zebra Added zebra handlers to 32-zebra Moved functions from 62-flyscans to others, see above Added question to 90-usersetup, should we use Chainmap?
