+++ startup/32-zebra.py
@@ -175,3 +175,372 @@ class SRXZebra(Zebra):
 
 zebra = SRXZebra('XF:05IDD-ES:1{Dev:Zebra1}:', name='zebra')
 zebra.read_attrs = ['pc.data.enc1', 'pc.data.time']
+
+class SRXFlyer1Axis(Device):
+#    LARGE_FILE_DIRECTORY_READ_PATH = '/tmp/test_data'
+#    LARGE_FILE_DIRECTORY_WRITE_PATH = '/tmp/fly_scan_ancillary'
+#     LARGE_FILE_DIRECTORY_WRITE_PATH = '/nsls2/xf05id1/XF05ID1/data/2018-1/fly_scan_ancillary/'
+#     LARGE_FILE_DIRECTORY_READ_PATH = '/nsls2/xf05id1/XF05ID1/data/2018-1/fly_scan_ancillary/'
+    LARGE_FILE_DIRECTORY_WRITE_PATH = '/nsls2/xf05id1/XF05ID1/data/2019-3/fly_scan_ancillary/'
+    LARGE_FILE_DIRECTORY_READ_PATH = '/nsls2/xf05id1/XF05ID1/data/2019-3/fly_scan_ancillary/'
+    "This is the Zebra."
+    KNOWN_DETS = {'xs', 'xs2', 'merlin', 'dexela'}
+    @property
+    def encoder(self):
+        return self._encoder
+
+    @property
+    def detectors(self):
+        return tuple(self._dets)
+
+    @detectors.setter
+    def detectors(self, value):
+        dets = tuple(value)
+        if not all(d.name in self.KNOWN_DETS
+                   for d in dets):
+            raise ValueError(f'One or more of {[d.name for d in dets]}'
+                             f'is not known to the zebara.  '
+                             f'The known detectors are {self.KNOWN_DETS})')
+        self._dets = dets
+
+    @property
+    def sclr(self):
+        return self._sis
+
+    def __init__(self, encoder, dets, sclr1, fast_axis, *, reg=db.reg, **kwargs):
+        super().__init__('', parent=None, **kwargs)
+        self._mode = 'idle'
+        self._encoder = encoder
+        self._dets = dets
+        self._sis = sclr1
+        self._filestore_resource = None
+
+        self._fast_axis = fast_axis
+
+        if self._fast_axis == 'HOR':
+            self.stage_sigs[self._encoder.pc.enc] = 'Enc2'
+            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
+            self.stage_sigs[self._encoder.pc.enc_res2] = 5E-6
+        elif self._fast_axis == 'VER':
+            self.stage_sigs[self._encoder.pc.enc] = 'Enc1'
+            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
+            self.stage_sigs[self._encoder.pc.enc_res1] = 5E-6
+        elif self._fast_axis == 'DET2HOR':
+            self.stage_sigs[self._encoder.pc.enc] = 'Enc3'
+            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
+            self.stage_sigs[self._encoder.pc.enc_res1] = 5E-5
+        elif self._fast_axis == 'DET2VER':
+            self.stage_sigs[self._encoder.pc.enc] = 'Enc4'
+            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
+            self.stage_sigs[self._encoder.pc.enc_res1] = 5E-5
+
+        # gating info for encoder capture
+        self.stage_sigs[self._encoder.pc.gate_num] = 1
+        self.stage_sigs[self._encoder.pc.pulse_start] = 0
+
+        #pc gate output is 31 for zebra.  use it to trigger xspress3 and I0
+        self.stage_sigs[self._encoder.output1.ttl.addr] = 31
+        self.stage_sigs[self._encoder.output3.ttl.addr] = 31
+        # this is for the merlin
+        self.stage_sigs[self._encoder.output2.ttl.addr] = 53
+        # this is for the dexela
+        # self.stage_sigs[self._encoder.output4.ttl.addr] = 55
+        # this is for the xs2
+        self.stage_sigs[self._encoder.output4.ttl.addr] = 31
+
+        self.stage_sigs[self._encoder.pc.enc_pos1_sync] = 1
+        self.stage_sigs[self._encoder.pc.enc_pos2_sync] = 1
+        self.stage_sigs[self._encoder.pc.enc_pos3_sync] = 1
+        self.stage_sigs[self._encoder.pc.enc_pos4_sync] = 1
+
+        #put SIS3820 into single count (not autocount) mode
+        self.stage_sigs[self._sis.count_mode] = 0
+
+        #stop the SIS3820
+        self._sis.stop_all.put(1)
+
+        self._encoder.pc.block_state_reset.put(1)
+        self.reg = reg
+        self._document_cache = []
+        self._last_bulk = None
+
+    def stage(self):
+        super().stage()
+
+    def describe_collect(self):
+
+        ext_spec = 'FileStore:'
+
+        spec = {'external': ext_spec,
+            'dtype' : 'array',
+            'shape' : [self._npts],
+            'source': ''  # make this the PV of the array the det is writing
+        }
+
+        desc = OrderedDict()
+        for chan in ('time', 'enc1'):
+            desc[chan] = spec
+            desc[chan]['source'] = getattr(self._encoder.pc.data, chan).pvname
+
+        # handle the detectors we are going to get
+        for d in self._dets:
+            desc.update(d.describe())
+
+        # handle the ion chamber that the zebra is collecting
+        desc['i0'] = spec
+        desc['i0']['source'] = self._sis.mca2.pvname
+        desc['i0_time'] = spec
+        desc['i0_time']['source'] = self._sis.mca1.pvname
+        desc['im'] = spec
+        desc['im']['source'] = self._sis.mca3.pvname
+        desc['it'] = spec
+        desc['it']['source'] = self._sis.mca4.pvname
+
+        return {'stream0': desc}
+
+
+    def kickoff(self, *, xstart, xstop, xnum, dwell):
+        dets_by_name = {d.name : d
+                        for d in self.detectors}
+
+        self._encoder.pc.arm.put(0)
+        self._mode = 'kicked off'
+        self._npts = int(xnum)
+        extent = xstop - xstart
+        pxsize = extent / (xnum-1)
+        # 2 ms delay between pulses
+        decrement = ((pxsize / dwell) * 0.002)
+        if decrement < 4e-6:
+            # print('Changing the pulse width')
+            decrement = 4e-6
+        self._encoder.pc.gate_start.put(xstart)
+        self._encoder.pc.gate_step.put(extent+0.0005)
+        self._encoder.pc.gate_width.put(extent+0.0001)
+
+        self._encoder.pc.pulse_start.put(0.0)
+        self._encoder.pc.pulse_max.put(xnum)
+        self._encoder.pc.pulse_step.put(pxsize)
+        self._encoder.pc.pulse_width.put(pxsize-decrement)
+        # If decrement is too small, then zebra will not send individual pulses
+        # but integrate over the entire line
+        # Hopefully taken care of with decrement check above
+
+        # For dexela, we will use time triggering in a pixel, not position
+        if 'dexela' in dets_by_name:
+            self._encoder.output1.ttl.addr.put(52)
+            self._encoder.output3.ttl.addr.put(52)
+            self._encoder.pulse1.width.put(0.5 * dwell - 0.050)
+        else:
+            self._encoder.output1.ttl.addr.put(31)
+            self._encoder.output3.ttl.addr.put(31)
+
+        # If both values are not synced, then the X-position was not updating
+        # during the scan and will remain at the initial value
+        # - AMK
+        self._encoder.pc.enc_pos1_sync.put(1)  # Sample Y
+        self._encoder.pc.enc_pos2_sync.put(1)  # Sample X
+        self._encoder.pc.enc_pos3_sync.put(1)  # Det2 Stage X
+        self._encoder.pc.enc_pos4_sync.put(1)  # Det2 Stage Y
+
+        # Arm the zebra
+        self._encoder.pc.arm.put(1)
+
+        st = NullStatus()  # TODO Return a status object *first* and do the above asynchronously.
+        return st
+
+    def complete(self):
+        """
+        Call this when all needed data has been collected. This has no idea
+        whether that is true, so it will obligingly stop immediately. It is
+        up to the caller to ensure that the motion is actually complete.
+        """
+        # Our acquisition complete PV is: XF:05IDD-ES:1{Dev:Zebra1}:ARRAY_ACQ
+        while self._encoder.pc.data_in_progress.get() == 1:
+            ttime.sleep(0.01)
+        # ttime.sleep(.1)
+        self._mode = 'complete'
+        self._encoder.pc.block_state_reset.put(1)
+        # see triggering errors of the xspress3 on suspension.  This is
+        # to test the reset of the xspress3 after a line.
+
+
+        for d in self._dets:
+            d.stop(success=True)
+
+        self.__filename = '{}.h5'.format(uuid.uuid4())
+        self.__filename_sis = '{}.h5'.format(uuid.uuid4())
+        self.__read_filepath = os.path.join(self.LARGE_FILE_DIRECTORY_READ_PATH,
+                                            self.__filename)
+        self.__read_filepath_sis = os.path.join(self.LARGE_FILE_DIRECTORY_READ_PATH,
+                                                self.__filename_sis)
+        self.__write_filepath = os.path.join(self.LARGE_FILE_DIRECTORY_WRITE_PATH,
+                                             self.__filename)
+        self.__write_filepath_sis = os.path.join(self.LARGE_FILE_DIRECTORY_WRITE_PATH,
+                                                 self.__filename_sis)
+
+        self.__filestore_resource, datum_factory_z = resource_factory(
+            'ZEBRA_HDF51', root='/',
+            resource_path=self.__read_filepath,
+            resource_kwargs={}, path_semantics='posix')
+        self.__filestore_resource_sis, datum_factory_sis = resource_factory(
+            'SIS_HDF51', root='/',
+            resource_path=self.__read_filepath_sis,
+            resource_kwargs={},
+            path_semantics='posix')
+
+        time_datum = datum_factory_z({'column': 'time'})
+        enc1_datum = datum_factory_z({'column': 'enc1'})
+        sis_datum =  datum_factory_sis({'column': 'i0'})
+        sis_datum_im =  datum_factory_sis({'column': 'im'})
+        sis_datum_it =  datum_factory_sis({'column': 'it'})
+        sis_time =  datum_factory_sis({'column': 'time'})
+
+        self._document_cache.extend(('resource', d) for d in (self.__filestore_resource,
+                                                             self.__filestore_resource_sis))
+        self._document_cache.extend(('datum', d) for d in (time_datum, enc1_datum,
+                                                          sis_datum, sis_time, sis_datum_im, sis_datum_it))
+
+        # grab the asset documents from all of the child detectors
+        for d in self._dets:
+            self._document_cache.extend(d.collect_asset_docs())
+
+        # Write the file.
+        # @timer_wrapper
+        def get_zebra_data():
+            export_zebra_data(self._encoder, self.__write_filepath,self._fast_axis)
+        # t_getzebradata = tic()
+        get_zebra_data()
+        # toc(t_getzebradata, str='Get Zebra data')
+
+        # @timer_wrapper
+        def get_sis_data():
+            export_sis_data(self._sis, self.__write_filepath_sis)
+        # t_sisdata = tic()
+        get_sis_data()
+        # toc(t_sisdata, str='Get SIS data')
+
+
+        # Yield a (partial) Event document. The RunEngine will put this
+        # into metadatastore, as it does all readings.
+        self._last_bulk =  {
+            'time': ttime.time(), 'seq_num': 1,
+            'data': {'time': time_datum['datum_id'],
+                     'enc1': enc1_datum['datum_id'],
+                     'i0': sis_datum['datum_id'],
+                     'i0_time': sis_time['datum_id'],
+                     'im': sis_datum_im['datum_id'],
+                     'it': sis_datum_it['datum_id']},
+            'timestamps': {'time': time_datum['datum_id'],  # not a typo#
+                           'enc1': time_datum['datum_id'],
+                           'i0': sis_time['datum_id'],
+                           'i0_time': sis_time['datum_id'],
+                           'im': sis_datum_im['datum_id'],
+                           'it': sis_datum_it['datum_id']}
+        }
+        for d in self._dets:
+            reading = d.read()
+            self._last_bulk['data'].update({k: v['value']
+                                            for k, v in reading.items()})
+            self._last_bulk['timestamps'].update({k: v['timestamp']
+                                                  for k, v in reading.items()})
+
+        return NullStatus()
+
+    def collect(self):
+        # Create records in the FileStore database.
+        # move this to stage because I thinkt hat describe_collect needs the
+        # resource id
+        # TODO use ophyd.areadectector.filestoer_mixins.resllource_factory here
+        if self._last_bulk is None:
+            raise Exception("the order of complete and collect is brittle and out "
+                            "of sync. This device relies on in-order and 1:1 calls "
+                            "between complete and collect to correctly create and stash "
+                            "the asset registry documents")
+        yield self._last_bulk
+        self._last_bulk = None
+        self._mode = 'idle'
+
+    def collect_asset_docs(self):
+        yield from iter(list(self._document_cache))
+        self._document_cache.clear()
+
+    def stop(self):
+        self._encoder.pc.block_state_reset.put(1)
+        pass
+
+    def pause(self):
+        "Pausing in the middle of a kickoff nukes the partial dataset."
+        self._encoder.pc.block_state_reset.put(1)
+        self._sis.stop_all.put(1)
+        for d in self._dets:
+            if hasattr(d, 'settings'):
+                d.settings.acquire.put(0)
+            if hasattr(d, 'cam'):
+                d.cam.acquire.put(0)
+        self._mode = 'idle'
+        self.unstage()
+
+    def resume(self):
+        self.unstage()
+        self.stage()
+
+flying_zebra = SRXFlyer1Axis(zebra, [xs], sclr1, 'HOR', name='flying_zebra')
+flying_zebra_y = SRXFlyer1Axis(zebra, [xs], sclr1, 'VER', name='flying_zebra_y')
+
+# For confocal
+if xs2 is not None:
+    flying_zebra_x_xs2 = SRXFlyer1Axis(zebra, [xs2], sclr1, 'HOR', name='flying_zebra_x_xs2')
+    flying_zebra_y_xs2 = SRXFlyer1Axis(zebra, [xs2], sclr1, 'VER', name='flying_zebra_y_xs2')
+else:
+    flying_zebra_x_xs2 = None
+    flying_zebra_y_xs2 = None
+# For chip imaging
+# flying_zebra_x_xs2 = SRXFlyer1Axis(zebra, xs2, sclr1, 'DET2HOR', name='flying_zebra')
+# flying_zebra_y_xs2 = SRXFlyer1Axis(zebra, xs2, sclr1, 'DET2VER', name='flying_zebra')
+# flying_zebra = SRXFlyer1Axis(zebra)
+
+
+def export_zebra_data(zebra, filepath, fast_axis):
+    j = 0
+    while zebra.pc.data_in_progress.get()==1:
+        print('waiting zebra')
+        ttime.sleep(0.1)
+        j += 1
+        if j > 10:
+            print('THE ZEBRA IS BEHAVING BADLY CARRYING ON')
+            break
+
+    time_d = zebra.pc.data.time.get()
+    if fast_axis == 'HOR':
+        enc1_d = zebra.pc.data.enc2.get()
+    elif fast_axis == 'DET2HOR':
+        enc1_d = zebra.pc.data.enc3.get()
+    elif fast_axis == 'DET2VER':
+        enc1_d = zebra.pc.data.enc4.get()
+    else:
+        enc1_d = zebra.pc.data.enc1.get()
+
+    while len(time_d) == 0 or len(time_d) != len(enc1_d):
+        time_d = zebra.pc.data.time.get()
+        if fast_axis == 'HOR':
+            enc1_d = zebra.pc.data.enc2.get()
+        else:
+            enc1_d = zebra.pc.data.enc1.get()
+
+    size = (len(time_d),)
+    with h5py.File(filepath, 'w') as f:
+        dset0 = f.create_dataset("time",size,dtype='f')
+        dset0[...] = np.array(time_d)
+        dset1 = f.create_dataset("enc1",size,dtype='f')
+        dset1[...] = np.array(enc1_d)
+        f.close()
+
+class ZebraHDF5Handler(HandlerBase):
+    HANDLER_NAME = 'ZEBRA_HDF51'
+    def __init__(self, resource_fn):
+        self._handle = h5py.File(resource_fn, 'r')
+
+    def __call__(self, *, column):
+        return self._handle[column][:]
+
+db.reg.register_handler('ZEBRA_HDF51', ZebraHDF5Handler, overwrite=True)
