+++ startup/62-flyscans.py
@@ -25,6 +25,7 @@ import uuid
 import h5py
 import numpy as np
 import time as ttime
+import matplotlib.pyplot as plt
 from collections import ChainMap
 
 from ophyd import Device
@@ -53,476 +54,14 @@ def timer_wrapper(func):
         print('%s: dt = %f' % (func.__name__, dt))
     return wrapper
 
+
 def tic():
     return ttime.monotonic()
+
 
 def toc(t0, str=''):
     dt = ttime.monotonic() - t0
     print('%s: dt = %f' % (str, dt))
-    return
-
-
-class SRXFlyer1Axis(Device):
-#    LARGE_FILE_DIRECTORY_READ_PATH = '/tmp/test_data'
-#    LARGE_FILE_DIRECTORY_WRITE_PATH = '/tmp/fly_scan_ancillary'
-#     LARGE_FILE_DIRECTORY_WRITE_PATH = '/nsls2/xf05id1/XF05ID1/data/2018-1/fly_scan_ancillary/'
-#     LARGE_FILE_DIRECTORY_READ_PATH = '/nsls2/xf05id1/XF05ID1/data/2018-1/fly_scan_ancillary/'
-    LARGE_FILE_DIRECTORY_WRITE_PATH = '/nsls2/xf05id1/XF05ID1/data/2019-3/fly_scan_ancillary/'
-    LARGE_FILE_DIRECTORY_READ_PATH = '/nsls2/xf05id1/XF05ID1/data/2019-3/fly_scan_ancillary/'
-    "This is the Zebra."
-    KNOWN_DETS = {'xs', 'xs2', 'merlin', 'dexela'}
-    @property
-    def encoder(self):
-        return self._encoder
-
-    @property
-    def detectors(self):
-        return tuple(self._dets)
-
-    @detectors.setter
-    def detectors(self, value):
-        dets = tuple(value)
-        if not all(d.name in self.KNOWN_DETS
-                   for d in dets):
-            raise ValueError(f'One or more of {[d.name for d in dets]}'
-                             f'is not known to the zebara.  '
-                             f'The known detectors are {self.KNOWN_DETS})')
-        self._dets = dets
-
-    @property
-    def sclr(self):
-        return self._sis
-
-    def __init__(self, encoder, dets, sclr1, fast_axis, *, reg=db.reg, **kwargs):
-        super().__init__('', parent=None, **kwargs)
-        self._mode = 'idle'
-        self._encoder = encoder
-        self._dets = dets
-        self._sis = sclr1
-        self._filestore_resource = None
-
-        self._fast_axis = fast_axis
-
-        if self._fast_axis == 'HOR':
-            self.stage_sigs[self._encoder.pc.enc] = 'Enc2'
-            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
-            self.stage_sigs[self._encoder.pc.enc_res2] = 5E-6
-        elif self._fast_axis == 'VER':
-            self.stage_sigs[self._encoder.pc.enc] = 'Enc1'
-            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
-            self.stage_sigs[self._encoder.pc.enc_res1] = 5E-6
-        elif self._fast_axis == 'DET2HOR':
-            self.stage_sigs[self._encoder.pc.enc] = 'Enc3'
-            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
-            self.stage_sigs[self._encoder.pc.enc_res1] = 5E-5
-        elif self._fast_axis == 'DET2VER':
-            self.stage_sigs[self._encoder.pc.enc] = 'Enc4'
-            self.stage_sigs[self._encoder.pc.dir] = 'Positive'
-            self.stage_sigs[self._encoder.pc.enc_res1] = 5E-5
-
-        # gating info for encoder capture
-        self.stage_sigs[self._encoder.pc.gate_num] = 1
-        self.stage_sigs[self._encoder.pc.pulse_start] = 0
-
-        #pc gate output is 31 for zebra.  use it to trigger xspress3 and I0
-        self.stage_sigs[self._encoder.output1.ttl.addr] = 31
-        self.stage_sigs[self._encoder.output3.ttl.addr] = 31
-        # this is for the merlin
-        self.stage_sigs[self._encoder.output2.ttl.addr] = 53
-        # this is for the dexela
-        # self.stage_sigs[self._encoder.output4.ttl.addr] = 55
-        # this is for the xs2
-        self.stage_sigs[self._encoder.output4.ttl.addr] = 31
-
-        self.stage_sigs[self._encoder.pc.enc_pos1_sync] = 1
-        self.stage_sigs[self._encoder.pc.enc_pos2_sync] = 1
-        self.stage_sigs[self._encoder.pc.enc_pos3_sync] = 1
-        self.stage_sigs[self._encoder.pc.enc_pos4_sync] = 1
-
-        #put SIS3820 into single count (not autocount) mode
-        self.stage_sigs[self._sis.count_mode] = 0
-
-        #stop the SIS3820
-        self._sis.stop_all.put(1)
-
-        self._encoder.pc.block_state_reset.put(1)
-        self.reg = reg
-        self._document_cache = []
-        self._last_bulk = None
-
-    def stage(self):
-        super().stage()
-
-    def describe_collect(self):
-
-        ext_spec = 'FileStore:'
-
-        spec = {'external': ext_spec,
-            'dtype' : 'array',
-            'shape' : [self._npts],
-            'source': ''  # make this the PV of the array the det is writing
-        }
-
-        desc = OrderedDict()
-        for chan in ('time', 'enc1'):
-            desc[chan] = spec
-            desc[chan]['source'] = getattr(self._encoder.pc.data, chan).pvname
-
-        # handle the detectors we are going to get
-        for d in self._dets:
-            desc.update(d.describe())
-
-        # handle the ion chamber that the zebra is collecting
-        desc['i0'] = spec
-        desc['i0']['source'] = self._sis.mca2.pvname
-        desc['i0_time'] = spec
-        desc['i0_time']['source'] = self._sis.mca1.pvname
-        desc['im'] = spec
-        desc['im']['source'] = self._sis.mca3.pvname
-        desc['it'] = spec
-        desc['it']['source'] = self._sis.mca4.pvname
-
-        return {'stream0': desc}
-
-
-    def kickoff(self, *, xstart, xstop, xnum, dwell):
-        dets_by_name = {d.name : d
-                        for d in self.detectors}
-
-        self._encoder.pc.arm.put(0)
-        self._mode = 'kicked off'
-        self._npts = int(xnum)
-        extent = xstop - xstart
-        pxsize = extent / (xnum-1)
-        # 2 ms delay between pulses
-        decrement = ((pxsize / dwell) * 0.002)
-        if decrement < 4e-6:
-            # print('Changing the pulse width')
-            decrement = 4e-6
-        self._encoder.pc.gate_start.put(xstart)
-        self._encoder.pc.gate_step.put(extent+0.0005)
-        self._encoder.pc.gate_width.put(extent+0.0001)
-
-        self._encoder.pc.pulse_start.put(0.0)
-        self._encoder.pc.pulse_max.put(xnum)
-        self._encoder.pc.pulse_step.put(pxsize)
-        self._encoder.pc.pulse_width.put(pxsize-decrement)
-        # If decrement is too small, then zebra will not send individual pulses
-        # but integrate over the entire line
-        # Hopefully taken care of with decrement check above
-
-        # For dexela, we will use time triggering in a pixel, not position
-        if 'dexela' in dets_by_name:
-            self._encoder.output1.ttl.addr.put(52)
-            self._encoder.output3.ttl.addr.put(52)
-            self._encoder.pulse1.width.put(0.5 * dwell - 0.050)
-        else:
-            self._encoder.output1.ttl.addr.put(31)
-            self._encoder.output3.ttl.addr.put(31)
-
-        # If both values are not synced, then the X-position was not updating
-        # during the scan and will remain at the initial value
-        # - AMK
-        self._encoder.pc.enc_pos1_sync.put(1)  # Sample Y
-        self._encoder.pc.enc_pos2_sync.put(1)  # Sample X
-        self._encoder.pc.enc_pos3_sync.put(1)  # Det2 Stage X
-        self._encoder.pc.enc_pos4_sync.put(1)  # Det2 Stage Y
-
-        # Arm the zebra
-        self._encoder.pc.arm.put(1)
-
-        st = NullStatus()  # TODO Return a status object *first* and do the above asynchronously.
-        return st
-
-    def complete(self):
-        """
-        Call this when all needed data has been collected. This has no idea
-        whether that is true, so it will obligingly stop immediately. It is
-        up to the caller to ensure that the motion is actually complete.
-        """
-        # Our acquisition complete PV is: XF:05IDD-ES:1{Dev:Zebra1}:ARRAY_ACQ
-        while self._encoder.pc.data_in_progress.get() == 1:
-            ttime.sleep(0.01)
-        # ttime.sleep(.1)
-        self._mode = 'complete'
-        self._encoder.pc.block_state_reset.put(1)
-        # see triggering errors of the xspress3 on suspension.  This is
-        # to test the reset of the xspress3 after a line.
-
-
-        for d in self._dets:
-            d.stop(success=True)
-
-        self.__filename = '{}.h5'.format(uuid.uuid4())
-        self.__filename_sis = '{}.h5'.format(uuid.uuid4())
-        self.__read_filepath = os.path.join(self.LARGE_FILE_DIRECTORY_READ_PATH,
-                                            self.__filename)
-        self.__read_filepath_sis = os.path.join(self.LARGE_FILE_DIRECTORY_READ_PATH,
-                                                self.__filename_sis)
-        self.__write_filepath = os.path.join(self.LARGE_FILE_DIRECTORY_WRITE_PATH,
-                                             self.__filename)
-        self.__write_filepath_sis = os.path.join(self.LARGE_FILE_DIRECTORY_WRITE_PATH,
-                                                 self.__filename_sis)
-
-        self.__filestore_resource, datum_factory_z = resource_factory(
-            'ZEBRA_HDF51', root='/',
-            resource_path=self.__read_filepath,
-            resource_kwargs={}, path_semantics='posix')
-        self.__filestore_resource_sis, datum_factory_sis = resource_factory(
-            'SIS_HDF51', root='/',
-            resource_path=self.__read_filepath_sis,
-            resource_kwargs={},
-            path_semantics='posix')
-
-        time_datum = datum_factory_z({'column': 'time'})
-        enc1_datum = datum_factory_z({'column': 'enc1'})
-        sis_datum =  datum_factory_sis({'column': 'i0'})
-        sis_datum_im =  datum_factory_sis({'column': 'im'})
-        sis_datum_it =  datum_factory_sis({'column': 'it'})
-        sis_time =  datum_factory_sis({'column': 'time'})
-
-        self._document_cache.extend(('resource', d) for d in (self.__filestore_resource,
-                                                             self.__filestore_resource_sis))
-        self._document_cache.extend(('datum', d) for d in (time_datum, enc1_datum,
-                                                          sis_datum, sis_time, sis_datum_im, sis_datum_it))
-
-        # grab the asset documents from all of the child detectors
-        for d in self._dets:
-            self._document_cache.extend(d.collect_asset_docs())
-
-        # Write the file.
-        # @timer_wrapper
-        def get_zebra_data():
-            export_zebra_data(self._encoder, self.__write_filepath,self._fast_axis)
-        # t_getzebradata = tic()
-        get_zebra_data()
-        # toc(t_getzebradata, str='Get Zebra data')
-
-        # @timer_wrapper
-        def get_sis_data():
-            export_sis_data(self._sis, self.__write_filepath_sis)
-        # t_sisdata = tic()
-        get_sis_data()
-        # toc(t_sisdata, str='Get SIS data')
-
-
-        # Yield a (partial) Event document. The RunEngine will put this
-        # into metadatastore, as it does all readings.
-        self._last_bulk =  {
-            'time': ttime.time(), 'seq_num': 1,
-            'data': {'time': time_datum['datum_id'],
-                     'enc1': enc1_datum['datum_id'],
-                     'i0': sis_datum['datum_id'],
-                     'i0_time': sis_time['datum_id'],
-                     'im': sis_datum_im['datum_id'],
-                     'it': sis_datum_it['datum_id']},
-            'timestamps': {'time': time_datum['datum_id'],  # not a typo#
-                           'enc1': time_datum['datum_id'],
-                           'i0': sis_time['datum_id'],
-                           'i0_time': sis_time['datum_id'],
-                           'im': sis_datum_im['datum_id'],
-                           'it': sis_datum_it['datum_id']}
-        }
-        for d in self._dets:
-            reading = d.read()
-            self._last_bulk['data'].update({k: v['value']
-                                            for k, v in reading.items()})
-            self._last_bulk['timestamps'].update({k: v['timestamp']
-                                                  for k, v in reading.items()})
-
-        return NullStatus()
-
-    def collect(self):
-        # Create records in the FileStore database.
-        # move this to stage because I thinkt hat describe_collect needs the
-        # resource id
-        # TODO use ophyd.areadectector.filestoer_mixins.resllource_factory here
-        if self._last_bulk is None:
-            raise Exception("the order of complete and collect is brittle and out "
-                            "of sync. This device relies on in-order and 1:1 calls "
-                            "between complete and collect to correctly create and stash "
-                            "the asset registry documents")
-        yield self._last_bulk
-        self._last_bulk = None
-        self._mode = 'idle'
-
-    def collect_asset_docs(self):
-        yield from iter(list(self._document_cache))
-        self._document_cache.clear()
-
-    def stop(self):
-        self._encoder.pc.block_state_reset.put(1)
-        pass
-
-    def pause(self):
-        "Pausing in the middle of a kickoff nukes the partial dataset."
-        self._encoder.pc.block_state_reset.put(1)
-        self._sis.stop_all.put(1)
-        for d in self._dets:
-            if hasattr(d, 'settings'):
-                d.settings.acquire.put(0)
-            if hasattr(d, 'cam'):
-                d.cam.acquire.put(0)
-        self._mode = 'idle'
-        self.unstage()
-
-    def resume(self):
-        self.unstage()
-        self.stage()
-       #  try:
-       #      self.stage()
-       #  except RedundantStaging:
-       #  # except:
-       #      print('Redundant Staging\nUnstaging flyer...')
-       #      self.unstage()
-       #      print('Staging flyer...')
-       #      self.stage()
-
-
-flying_zebra = SRXFlyer1Axis(zebra, [xs], sclr1, 'HOR', name='flying_zebra')
-flying_zebra_y = SRXFlyer1Axis(zebra, [xs], sclr1, 'VER', name='flying_zebra_y')
-
-# For confocal
-if xs2 is not None:
-    flying_zebra_x_xs2 = SRXFlyer1Axis(zebra, [xs2], sclr1, 'HOR', name='flying_zebra_x_xs2')
-    flying_zebra_y_xs2 = SRXFlyer1Axis(zebra, [xs2], sclr1, 'VER', name='flying_zebra_y_xs2')
-else:
-    flying_zebra_x_xs2 = None
-    flying_zebra_y_xs2 = None
-# For chip imaging
-# flying_zebra_x_xs2 = SRXFlyer1Axis(zebra, xs2, sclr1, 'DET2HOR', name='flying_zebra')
-# flying_zebra_y_xs2 = SRXFlyer1Axis(zebra, xs2, sclr1, 'DET2VER', name='flying_zebra')
-# flying_zebra = SRXFlyer1Axis(zebra)
-
-
-def export_zebra_data(zebra, filepath, fast_axis):
-    j = 0
-    while zebra.pc.data_in_progress.get()==1:
-        print('waiting zebra')
-        ttime.sleep(0.1)
-        j += 1
-        if j > 10:
-            print('THE ZEBRA IS BEHAVING BADLY CARRYING ON')
-            break
-
-    time_d = zebra.pc.data.time.get()
-    if fast_axis == 'HOR':
-        enc1_d = zebra.pc.data.enc2.get()
-    elif fast_axis == 'DET2HOR':
-        enc1_d = zebra.pc.data.enc3.get()
-    elif fast_axis == 'DET2VER':
-        enc1_d = zebra.pc.data.enc4.get()
-    else:
-        enc1_d = zebra.pc.data.enc1.get()
-
-    while len(time_d) == 0 or len(time_d) != len(enc1_d):
-        time_d = zebra.pc.data.time.get()
-        if fast_axis == 'HOR':
-            enc1_d = zebra.pc.data.enc2.get()
-        else:
-            enc1_d = zebra.pc.data.enc1.get()
-
-    size = (len(time_d),)
-    with h5py.File(filepath, 'w') as f:
-        dset0 = f.create_dataset("time",size,dtype='f')
-        dset0[...] = np.array(time_d)
-        dset1 = f.create_dataset("enc1",size,dtype='f')
-        dset1[...] = np.array(enc1_d)
-        f.close()
-
-def export_sis_data(ion, filepath):
-    t = ion.mca1.get(timeout=5.)
-    i = ion.mca2.get(timeout=5.)
-    im= ion.mca3.get(timeout=5.)
-    it= ion.mca4.get(timeout=5.)
-    while len(t) == 0 and len(t) != len(i):
-        t = ion.mca1.get(timeout=5.)
-        i = ion.mca2.get(timeout=5.)
-        im= ion.mca3.get(timeout=5.)
-        it= ion.mca4.get(timeout=5.)
-
-    correct_length = zebra.pc.data.num_down.get()
-    # size = (len(t),)
-    # size2 = (len(i),)
-    # size3 = (len(im),)
-    # size4 = (len(it),)
-    with h5py.File(filepath, 'w') as f:
-        if len(t) != correct_length:
-            correction_factor = (correct_length-len(t))
-            new_t = [k for k in t] + [ 1e10 for _ in range(0,int(correction_factor)) ]
-            new_i = [k for k in i] + [ 1e10 for _ in range(0,int(correction_factor)) ]
-            new_im= [k for k in im] + [ 1e10 for _ in range(0,int(correction_factor)) ]
-            new_it= [k for k in it] + [ 1e10 for _ in range(0,int(correction_factor)) ]
-        else:
-            correction_factor = 0
-            new_t = t
-            new_i = i
-            new_im= im
-            new_it= it
-        dset0 = f.create_dataset("time",(correct_length,),dtype='f')
-        dset0[...] = np.array(new_t)
-        dset1 = f.create_dataset("i0",(correct_length,),dtype='f')
-        dset1[...] = np.array(new_i)
-        dset2 = f.create_dataset("im",(correct_length,),dtype='f')
-        dset2[...] = np.array(new_im)
-        dset3 = f.create_dataset("it",(correct_length,),dtype='f')
-        dset3[...] = np.array(new_it)
-        f.close()
-
-class ZebraHDF5Handler(HandlerBase):
-    HANDLER_NAME = 'ZEBRA_HDF51'
-    def __init__(self, resource_fn):
-        self._handle = h5py.File(resource_fn, 'r')
-
-    def __call__(self, *, column):
-        return self._handle[column][:]
-
-class SISHDF5Handler(HandlerBase):
-    HANDLER_NAME = 'SIS_HDF51'
-    def __init__(self, resource_fn):
-        self._handle = h5py.File(resource_fn, 'r')
-
-    def __call__(self, *, column):
-        return self._handle[column][:]
-
-
-db.reg.register_handler('SIS_HDF51', SISHDF5Handler, overwrite=True)
-db.reg.register_handler('ZEBRA_HDF51', ZebraHDF5Handler, overwrite=True)
-
-
-class LiveZebraPlot(CallbackBase):
-    """
-    This is a really dumb approach but it gets the job done. To fix later.
-    """
-
-    def __init__(self, ax=None):
-        self._uid = None
-        self._desc_uid = None
-        if ax is None:
-            fig, ax = plt.subplots()
-        self.ax = ax
-        self.legend_title = 'sequence #'
-
-    def start(self, doc):
-        self._uid = doc['uid']
-
-    def descriptor(self, doc):
-        if doc['name'] == 'stream0':
-            self._desc_uid = doc['uid']
-
-    def bulk_events(self, docs):
-        # Don't actually use the docs, but use the fact that they have been
-        # emitted as a signal to go grab the data from the databroker now.
-        event_uids = [doc['uid'] for doc in docs[self._desc_uid]]
-        events = db.get_events(db[self._uid], stream_name='stream0', fields=['enc1', 'time'], fill=True)
-        for event in events:
-            if event['uid'] in event_uids:
-                self.ax.plot(event['data']['time'], event['data']['enc1'], label=event['seq_num'])
-        self.ax.legend(loc=0, title=self.legend_title)
-
-    def stop(self, doc):
-        self._uid = None
-        self._desc_uid = None
 
 
 # changed the flyer device to be aware of fast vs slow axis in a 2D scan
@@ -610,7 +149,7 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
         dpc.hdf5.stage_sigs['num_capture'] = xnum
         del dpc
 
-    # setup dexela
+    # Setup dexela
     if ('dexela' in dets_by_name):
         xrd = dets_by_name['dexela']
         xrd.cam.stage_sigs['acquire_time'] = 0.50 * dwell - 0.050
@@ -618,7 +157,7 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
         del xrd
 
     # If delta is None, set delta based on time for acceleration
-    if delta is None:
+    if (delta is None):
         MIN_DELTA = 0.002  # old default value
         v = ((xstop - xstart) / (xnum-1)) / dwell  # compute "stage speed"
         t_acc = 1.0  # acceleration time, default 1.0 s
@@ -626,11 +165,12 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
         delta = np.amax((delta, MIN_DELTA))
 
     # Move to start scanning location
-    yield from mv(xmotor, xstart - delta, ymotor, ystart)
+    yield from mv(xmotor, xstart - delta,
+                  ymotor, ystart)
 
     # Run a peakup before the map?
-    if align:
-        yield from peakup_fine()
+    if (align):
+        yield from peakup_fine(shutter=shutter)
 
     md = ChainMap(md, {
         'plan_name': 'scan_and_fly',
@@ -641,9 +181,6 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
                      'raster': False,
                      'fast_axis': flying_zebra._fast_axis},
                      'theta': hf_stage.th.position,
-        # 'scaninfo': {'type': 'E_tomo',
-        #              'raster': False,
-        #              'fast_axis': flying_zebra._fast_axis},
         'scan_params': [xstart, xstop, xnum, ystart, ystop, ynum, dwell],
         'scan_input': [xstart, xstop, xnum, ystart, ystop, ynum, dwell],
         'delta': delta,
@@ -658,11 +195,6 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
     def fly_each_step(motor, step):
         def move_to_start_fly():
             "See http://nsls-ii.github.io/bluesky/plans.html#the-per-step-hook"
-            # First, let 'scan' handle the normal y step, including a checkpoint.
-            # yield from one_nd_step([], {xmotor : xstart - delta,
-            #                             ymotor : step},
-            #                            {xmotor : xmotor.position,
-            #                             ymotor : ymotor.position})
             yield from abs_set(xmotor, xstart-delta, group='row')
             yield from one_1d_step([], motor, step)
             yield from bps.wait(group='row')
@@ -697,26 +229,26 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
 
         # set up all of the detectors
         # TODO we should be able to move this out of the per-line call?!
-        if 'xs' in dets_by_name:
+        if ('xs' in dets_by_name):
             xs = dets_by_name['xs']
             # yield from abs_set(xs.hdf5.num_capture, xnum, wait=True)
             # yield from abs_set(xs.settings.num_images, xnum, wait=True)
             yield from mv(xs.hdf5.num_capture, xnum,
                           xs.settings.num_images, xnum)
 
-        if 'xs2' in dets_by_name:
+        if ('xs2' in dets_by_name):
             xs2 = dets_by_name['xs2']
             # yield from abs_set(xs2.hdf5.num_capture, xnum, wait=True)
             # yield from abs_set(xs2.settings.num_images, xnum, wait=True)
             yield from mv(xs2.hdf5.num_capture, xnum,
                           xs2.settings.num_images, xnum)
 
-        if 'merlin' in dets_by_name:
+        if ('merlin' in dets_by_name):
             merlin = dets_by_name['merlin']
             yield from abs_set(merlin.hdf5.num_capture, xnum, wait=True)
             yield from abs_set(merlin.cam.num_images, xnum, wait=True)
 
-        if 'dexela' in dets_by_name:
+        if ('dexela' in dets_by_name):
             dexela = dets_by_name['dexela']
             yield from abs_set(dexela.hdf5.num_capture, xnum, wait=True)
             yield from abs_set(dexela.cam.num_images, xnum, wait=True)
@@ -728,8 +260,8 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
         # @timer_wrapper
         def zebra_kickoff():
             yield from kickoff(flying_zebra,
-                           xstart=xstart, xstop=xstop, xnum=xnum, dwell=dwell,
-                           wait=True)
+                               xstart=xstart, xstop=xstop, xnum=xnum, dwell=dwell,
+                               wait=True)
         # t_zebkickoff = tic()
         yield from zebra_kickoff()
         # toc(t_zebkickoff, str='Zebra kickoff')
@@ -742,7 +274,7 @@ def scan_and_fly_base(detectors, xstart, xstop, xnum, ystart, ystop, ynum, dwell
         # trigger all of the detectors
         for d in flying_zebra.detectors:
             yield from bps.trigger(d, group='row')
-            if d.name == 'dexela':
+            if (d.name == 'dexela'):
                 yield from bps.sleep(1)
 
         yield from bps.sleep(1.5)
@@ -879,6 +411,64 @@ def scan_and_fly(*args, extra_dets=None, **kwargs):
     yield from scan_and_fly_base(dets, *args, **kwargs)
 
 
+def y_scan_and_fly(*args, extra_dets=None, **kwargs):
+    '''
+    Convenience wrapper for scanning Y as the fast axis.
+    Call scan_and_fly_base, forcing slow and fast axes to be X and Y.
+    In this function, the first three scan parameters are for the *fast axis*,
+    i.e., the vertical, and the second three for the *slow axis*, horizontal.
+    '''
+
+    kwargs.setdefault('xmotor', hf_stage.y)
+    kwargs.setdefault('ymotor', hf_stage.x)
+    kwargs.setdefault('flying_zebra', flying_zebra_y)
+
+    _xs = kwargs.pop('xs', xs)
+    if (extra_dets is None):
+        extra_dets = []
+    dets = [_xs] + extra_dets
+    yield from scan_and_fly_base(dets, *args, **kwargs)
+
+
+# This class is not used in this file
+class LiveZebraPlot(CallbackBase):
+    """
+    This is a really dumb approach but it gets the job done. To fix later.
+    """
+
+    def __init__(self, ax=None):
+        self._uid = None
+        self._desc_uid = None
+        if (ax is None):
+            _, ax = plt.subplots()
+        self.ax = ax
+        self.legend_title = 'sequence #'
+
+    def start(self, doc):
+        self._uid = doc['uid']
+
+    def descriptor(self, doc):
+        if doc['name'] == 'stream0':
+            self._desc_uid = doc['uid']
+
+    def bulk_events(self, docs):
+        # Don't actually use the docs, but use the fact that they have been
+        # emitted as a signal to go grab the data from the databroker now.
+        event_uids = [doc['uid'] for doc in docs[self._desc_uid]]
+        events = db.get_events(db[self._uid], stream_name='stream0',
+                               fields=['enc1', 'time'], fill=True)
+        for event in events:
+            if (event['uid'] in event_uids):
+                self.ax.plot(event['data']['time'], event['data']['enc1'],
+                             label=event['seq_num'])
+        self.ax.legend(loc=0, title=self.legend_title)
+
+    def stop(self, doc):
+        self._uid = None
+        self._desc_uid = None
+
+
+# This class is not used in this file
 class RowBasedLiveGrid(LiveGrid):
     """
     Synthesize info from two event stream here.
@@ -929,77 +519,9 @@ class RowBasedLiveGrid(LiveGrid):
         self._pad = None
 
 
-class SrxXSP3Handler:
-    XRF_DATA_KEY = 'entry/instrument/detector/data'
-
-    def __init__(self, filepath, **kwargs):
-        self._filepath = filepath
-
-    def __call__(self, **kwargs):
-        with h5py.File(self._filepath, 'r') as f:
-            return np.asarray(f[self.XRF_DATA_KEY])
-
-
-def batch_fly(paramlist, kwlist=None, zlist=None):
-    '''
-    paramlist   list    list of positional and dwell time arguments to scan_and_fly
-    kwlist      list    list of dicts containing keywords to pass to scan_and_fly
-    '''
-
-    if kwlist == None:
-        kwlist=list()
-        for _ in paramlist:
-            kwlist.append({})
-
-    for i in range(0,len(paramlist)):
-        #this should be made more general
-        if zlist is not None:
-            yield from abs_set(hf_stage.z,zlist[i],wait=True)
-        yield from scan_and_fly(*paramlist[i],**kwlist[i])
-
-
-def batch_fly_arb(paramlist, kwlist=None, motlist=None):
-    '''
-    paramlist   list            list of positional and dwell time arguments to scan_and_fly
-    kwlist      list            list of dicts containing keywords to pass to scan_and_fly
-    motlist     list of lists   a list of motor,value pairs that define the prestart condition
-    '''
-
-    if kwlist is None:
-        kwlist = list()
-        for _ in paramlist:
-            kwlist.append({})
-
-#    for i in range(0,len(paramlist)):
-#        if motlist is not None:
-#            yield from abs_set(motlist[i][0],motlist[i][1],wait=True)
-#        yield from scan_and_fly(*paramlist[i],**kwlist[i])
-    for i in range(0, len(paramlist)):
-        if motlist is not None:
-            for pospair in motlist[i]:
-                yield from abs_set(pospair[0], pospair[1], wait=True)
-        yield from scan_and_fly(*paramlist[i], **kwlist[i])
-
-
-def y_scan_and_fly(*args, extra_dets=None, **kwargs):
-    '''
-    Convenience wrapper for scanning Y as the fast axis.
-    Call scan_and_fly_base, forcing slow and fast axes to be X and Y.
-    In this function, the first three scan parameters are for the *fast axis*,
-    i.e., the vertical, and the second three for the *slow axis*, horizontal.
-    '''
-
-    kwargs.setdefault('xmotor', hf_stage.y)
-    kwargs.setdefault('ymotor', hf_stage.x)
-    kwargs.setdefault('flying_zebra', flying_zebra_y)
-
-    _xs = kwargs.pop('xs', xs)
-    if extra_dets is None:
-        extra_dets = []
-    dets = [_xs] + extra_dets
-    yield from scan_and_fly_base(dets, *args, **kwargs)
-
-
+###############################################################################################
+# Convenience wrappers for different scans
+# This can probably move to a "user scans" file
 def y_scan_and_fly_xs2(*args, extra_dets=None, **kwargs):
     '''
     Convenience wrapper for scanning Y as the fast axis.
